refactor entities per exchange hub
