+++ src/entities/index.ts
@@ -1,7 +1,12 @@
-export { Asset } from './Asset';
-export { Instrument } from './Instrument';
-export { Trade } from './Trade';
-export { OrderBook as Book } from './OrderBook';
-export { Order } from './Order';
-export { Wallet } from './Wallet';
-export { Position } from './Position';
+export { createEntities } from './createEntities';
+export type {
+    Entities,
+    EntityClass,
+    Asset,
+    Instrument,
+    OrderBook,
+    Order,
+    Trade,
+    Wallet,
+    Position,
+} from './createEntities';
